function to calculate distance to plane with signs
Add a new function that can calculate distance to a fitting plane with signs. The distance from the given reference coordinate to the plane is positive.
Output parts with progress bar are rewritten using new class ProgressBar.
Opencv-python module cv2 would not be imported at the beginning of loading.
Fix an error in gitignore file.